Move default_argument_parser to smarts.core.utils
