+++ examples/dynamic_history_vehicles_replacement.py
@@ -3,7 +3,6 @@ import random
 from typing import Any, Callable, Dict, Sequence
 
 from envision.client import Client as Envision
-from examples.argument_parser import default_argument_parser
 from smarts.core import seed as random_seed
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface
@@ -11,6 +10,7 @@ from smarts.core.controllers import ActionSpaceType
 from smarts.core.plan import PlanningError
 from smarts.core.scenario import Scenario
 from smarts.core.smarts import SMARTS
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.zoo.agent_spec import AgentSpec
 
 logging.basicConfig(level=logging.INFO)
