+++ examples/egoless.py
@@ -1,8 +1,7 @@
 import logging
 
 import gym
-
-from examples.argument_parser import default_argument_parser
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
 
 logging.basicConfig(level=logging.INFO)
