+++ examples/multi_agent.py
@@ -1,11 +1,10 @@
 import pathlib
 
 import gym
-
-from examples.argument_parser import default_argument_parser
 from smarts import sstudio
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface, AgentType
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
 from smarts.zoo.agent_spec import AgentSpec
