+++ examples/parallel_environment.py
@@ -7,11 +7,11 @@ gym.logger.set_level(40)
 from functools import partial
 from typing import Dict, Sequence, Tuple
 
-from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface
 from smarts.core.controllers import ActionSpaceType
 from smarts.core.sensors import Observation
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.env.hiway_env import HiWayEnv
 from smarts.env.wrappers.frame_stack import FrameStack
 from smarts.env.wrappers.parallel_env import ParallelEnv
