+++ examples/ray_multi_instance.py
@@ -17,9 +17,9 @@ except Exception as e:
     raise RayException.required_to("ray_multi_instance.py")
 
 
-from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface, AgentType
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
 from smarts.zoo.agent_spec import AgentSpec
