+++ examples/replay/replay_klws_agent.py
@@ -5,8 +5,7 @@ import stat
 import sys
 
 import gym
-
-from examples.argument_parser import default_argument_parser
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
 from smarts.zoo.registry import make as zoo_make
