+++ examples/single_agent.py
@@ -2,12 +2,11 @@ import logging
 import pathlib
 
 import gym
-
-from examples.argument_parser import default_argument_parser
 from smarts import sstudio
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.sensors import Observation
+from smarts.core.utils.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
 from smarts.env.wrappers.single_agent import SingleAgent
 from smarts.zoo.agent_spec import AgentSpec
