+++ examples/traffic_histories_to_observations.py
@@ -5,9 +5,8 @@ import pickle
 from dataclasses import replace
 from typing import Optional, Sequence
 
-from PIL import Image, ImageDraw
-
 from envision.client import Client as Envision
+from PIL import Image, ImageDraw
 from smarts import sstudio
 from smarts.core import seed as smarts_seed
 from smarts.core.agent_interface import (
